s3-printing: fix crash bug in print_spool_open().
Guenther

Autobuild-User: Günther Deschner <gd@samba.org>
Autobuild-Date: Wed Dec 22 14:21:58 CET 2010 on sn-devel-104
(cherry picked from commit 94e83e385e5ba66f1415164214f7afd5e4ad7536)

+++ source3/printing/printspoolss.c
@@ -208,7 +208,9 @@ done:
 	if (!NT_STATUS_IS_OK(status)) {
 		if (fd != -1) {
 			close(fd);
-			unlink(fsp->print_file->filename);
+			if (fsp->print_file) {
+				unlink(fsp->print_file->filename);
+			}
 		}
 		/* We need to delete the job from spoolss too */
 		if (pf->jobid) {
